Muutin +page sivun ja vaihoin tiedot kysymys komponenttiin
