+++ svelte/src/lib/components/Kysymys.svelte
@@ -18,37 +18,34 @@
 			throw new Error('Dataa ei löytynyt');
 		}
 		kysymykset = await response.json();
+		randomKysymykset(kysymykset);
 	});
 
 	let valitutKysymykset: Kysymys[] = [];
 	function randomKysymykset(taulukko: Kysymys[]) {
-		//Virheen tarkistus tähän
-
 		//Tämä määrää, montako kysymystä halutaan
 		const montaKysymysta = 5;
-		for (let i = 0; i <= montaKysymysta; i++) {
+		//Virheen tarkistus tähän
+		if (taulukko.length < montaKysymysta) {
+			throw new Error('Data liian pieni');
+		}
+		while (valitutKysymykset.length < montaKysymysta) {
 			//Tähän muuttujaan tallennetaan hetkellisesti kysymys
 			let a: Kysymys;
 
 			//random numero 0-taulukonpituus
 			//virheen tarkistus?
-			const ind: number = Math.floor(Math.random() * kysymykset.length);
+			const ind: number = Math.floor(Math.random() * taulukko.length);
 
 			//Tässä tallennetaan muuttujaan taulukosta kyseisen indeksin tiedot
-			a = kysymykset[ind];
+			a = taulukko[ind];
 
 			//Käydään läpi valitutkysymykset taulukko, onko kyseinen objekti jo taulukossa (jos on, älä tee mitään, jos ei ole, push), tämä tehdään id:n avulla
-			//Jos taulukossa ei vielä ole mitään, pushaa
-			if (valitutKysymykset.length === 0) {
-				valitutKysymykset.push(a);
-				continue;
-			}
-
-			if (!valitutKysymykset.some((onkoID) => onkoID.id === a.id)) {
+			if (!valitutKysymykset.some((onkoID) => a.id === onkoID.id)) {
 				valitutKysymykset.push(a);
 			}
 		}
+		console.log(valitutKysymykset);
 		return valitutKysymykset;
 	}
-	console.log(valitutKysymykset);
 </script>
